Moved middleware container functions to container module
